Update middleware dependencies and apply type changes
This commit updates the middleware dependencies, replacing the Fideloper and Fruitcake packages with their Illuminate counterparts. It adjusts the types in TrustProxies and DatabaseHandler files to enhance the type-safety. In the composer.json file, outdated packages have been removed and remaining ones have been updated to their latest stable versions.