Remove non-breaking spaces from content and description
The changes allow removal of non-breaking spaces from the 'description' and 'content' fields by using str_replace function. This ensures clean and uniform data storage for these fields in the EpisodesRepository and ShowsRepository. Additionally, updated the 'findById' function in ShowsRepository to order by id in ascending order for a more organized data return.