Update support for YouTube and extend the copyright date
This commit replaces "Google Podcast" with "YouTube" in the description for our podcast services, reflecting that we now support YouTube as a platform for our users. Moreover, the copyright date in footer has been extended to 2024. The image in the docker-compose file has also been updated to "kuchura/your-cast-frontend:1.0.1-8df2a01".