Add new Geo and Agent integrations and update docker setup
Introduced new integrations to manage geo-coding and agent parsing in the application. The new classes integrate with external services to gather geo-information from an IP address and determine the platform of a user from the user agent string. The docker setup also faced changes with updates to the nginx and Postgres Dockerfile, along with the docker-compose configuration. Also made minor changes in certain PHP classes and configurations removing extra whitespace and fixing coding standards.